+++ src/fetch_agent/setup.py
@@ -10,6 +10,7 @@ setup(
         ('share/ament_index/resource_index/packages',
          ['resource/' + package_name]),
         ('share/' + package_name, ['package.xml']),
+        ('lib/' + package_name, [package_name + '/communication.py', package_name + '/agent.py', package_name + '/fetchAgent.py']),
     ],
     install_requires=['setuptools'],
     zip_safe=True,
